Added GID locator registration filtering. (#4)
This update ships an automatic local application name (+app_name+)
detection (for Rails applications only at the moment). That change is
the foundation for the possibility to skip GID locator registration with
the same name as the local application. For regular users this should
rarely be an issue, but for Hausgold internal apps this may cause
conflicts.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>