Implemented the ability to search entities. (#11)
This update ships the implementation of the query/search functionality.
This features low-level HTTP client APIs, client driver APIs and a brand
new search package which consists of three parts. The query settings, to
frame the search indention. The paging component which takes care of
planned page layouts and request optimizations. And last but not least a
search executor which serves the Ruby Enumerable API by combining the
query settings and paging details into actual client driver actions.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>