Implemented a Verkaeuferportal API client. (#25)
This update ships a client implementation for the
Verkaeuferportal/Kundenportal API (Grape migration). With support for
the Kundenportal users (customers). To make this work a bunch of
enhancements was added to support application and entity name aliases on
Global Ids, both for locating and writing directions. This patch also
ships example documentations how to use the new `Hausgold::Customer`.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>